New Payed Fees Checker and Input for Hourly Pay
The Payed Fees Checker addition to this program ensures that the net pay displayed to the user is always positive but if the fees weren't fully paid, they will be listed as still being owed. The new hourly rate input replaced the constant $16.78 by swapping the constant HOURRATE with a variable inputRate that is prompted for from the user. The inputRate is also ensured to always be positive via a while loop addition that checks it is $0.00 or greater and prompts the user for a new hourly rate if that isn't the case.